move enum
