+++ playwright-e2e/dsm/pages/participant-page/enums/download-format-enum.ts
@@ -1,0 +1,9 @@
+export enum FileFormatEnum {
+  XLSX = 'Excel (.xlsx)',
+  TSV = 'Tab-delimited (.tsv)'
+}
+
+export enum TextFormatEnum {
+  HUMAN_READABLE = 'Human-readable',
+  ANALYSIS_FRIENDLY = 'Analysis-friendly'
+}
