+++ playwright-e2e/tests/dsm/participant-list/download-subset-data.spec.ts
@@ -1,6 +1,7 @@
 import { expect } from '@playwright/test';
 import { StudyEnum } from 'dsm/component/navigation/enums/selectStudyNav-enum';
-import ParticipantListPage, { FileFormatEnum, TextFormatEnum } from 'dsm/pages/participant-list-page';
+import ParticipantListPage from 'dsm/pages/participant-list-page';
+import { FileFormatEnum, TextFormatEnum } from 'dsm/pages/participant-page/enums/download-format-enum';
 import { test } from 'fixtures/dsm-fixture';
 import { assertParticipantListDownloadFileName } from 'utils/test-utils';
